Move tools to a subdirectory

+++ tools/build.sh
@@ -6,6 +6,10 @@ if [[ ! -f logo.svg ]]; then
   echo >&2 "Unable to find Knola sources"
   exit 1
 fi
+if [[ ! -x tools/marked ]]; then
+  echo >&2 "Unable to find marked tool"
+  exit 1
+fi
 
 rm -rf _public
 mkdir -p _build _public _public/_resources
@@ -14,7 +18,7 @@ for file in *.md; do
   base=${file%.md}
   doc=_build/${base}.html
   cat header.html > $doc
-  ./marked $file >> $doc
+  tools/marked $file >> $doc
   cat footer.html >> $doc
   ln $doc _public/$base 
 done
